requirements.txt
ffmpeg-python
pillow
requests
boto3
supabase
yt-dlp
python-dotenv